refactor: extract video rendering pipeline into minarch_video.c
Moves bitmap_font, drawRect, fillRect, blitBitmapText, drawGauge,
selectScaler, screen_flip, applyFadeIn, drawDebugHud,
video_refresh_callback_main, convert_xrgb8888_to_rgba,
convert_rgb565_to_rgba, and video_refresh_callback out of minarch.c.

Exposes selectScaler and video_refresh_callback as public API (needed
by Core_open and the menu scaling path respectively). Promotes
ambient_mode, show_debug, DEVICE_PITCH, and shader_reset_suppressed
to non-static so the video module can access them. Adds Video_cleanup()
to safely free the RGBA conversion buffer.
